server/requirements.txt
flask
flask-cors
opencv-python-headless
numpy
Pillow
tensorflow
uuid
python-dotenv
psycopg2-binary
